terminal: improve default font and Nerd symbols fallback
