+++ src/browser/components/TerminalView.tsx
@@ -8,6 +8,15 @@ import {
   type TerminalFontConfig,
 } from "@/common/constants/storage";
 import { useTerminalRouter } from "@/browser/terminal/TerminalRouterContext";
+import {
+  appendTerminalIconFallback,
+  formatCssFontFamilyList,
+  isFontFamilyAvailableInBrowser,
+  isGenericFontFamily,
+  splitFontFamilyList,
+  stripOuterQuotes,
+  TERMINAL_ICON_FALLBACK_FAMILY,
+} from "@/browser/terminal/terminalFontFamily";
 import { TERMINAL_CONTAINER_ATTR } from "@/browser/utils/ui/keybinds";
 
 function normalizeTerminalFontConfig(value: unknown): TerminalFontConfig {
@@ -29,148 +38,6 @@ function normalizeTerminalFontConfig(value: unknown): TerminalFontConfig {
       : DEFAULT_TERMINAL_FONT_CONFIG.fontSize;
 
   return { fontFamily, fontSize };
-}
-
-const GENERIC_FONT_FAMILIES: ReadonlySet<string> = new Set([
-  "serif",
-  "sans-serif",
-  "monospace",
-  "cursive",
-  "fantasy",
-  "system-ui",
-  "ui-serif",
-  "ui-sans-serif",
-  "ui-monospace",
-  "ui-rounded",
-  "emoji",
-  "math",
-  "fangsong",
-]);
-
-function stripOuterQuotes(value: string): string {
-  const trimmed = value.trim();
-  if (
-    (trimmed.startsWith('"') && trimmed.endsWith('"')) ||
-    (trimmed.startsWith("'") && trimmed.endsWith("'"))
-  ) {
-    return trimmed.slice(1, -1);
-  }
-
-  return trimmed;
-}
-
-function splitFontFamilyList(value: string): string[] {
-  return value
-    .split(",")
-    .map((part) => part.trim())
-    .filter(Boolean);
-}
-
-function isGenericFontFamily(value: string): boolean {
-  return GENERIC_FONT_FAMILIES.has(value.trim().toLowerCase());
-}
-
-function quoteCssFontFamily(value: string): string {
-  const trimmed = stripOuterQuotes(value);
-  if (!trimmed) {
-    return trimmed;
-  }
-
-  if (isGenericFontFamily(trimmed)) {
-    return trimmed;
-  }
-
-  // Quote names containing whitespace or punctuation that would confuse CSS parsing.
-  //
-  // If the name itself contains quotes, strip them. A font-family name containing literal
-  // quotes is almost certainly user input error and would produce invalid CSS.
-  if (/[^a-zA-Z0-9_-]/.test(trimmed)) {
-    const sanitized = trimmed.replace(/"/g, "");
-    return `"${sanitized}"`;
-  }
-
-  return trimmed;
-}
-
-function formatCssFontFamilyList(value: string): string {
-  const parts = splitFontFamilyList(value);
-  if (parts.length === 0) {
-    return value.trim();
-  }
-
-  return parts.map(quoteCssFontFamily).join(", ");
-}
-
-const TERMINAL_ICON_FALLBACK_FAMILY = "Nerd Font Symbols";
-
-function appendTerminalIconFallback(fontFamily: string): string {
-  const parts = splitFontFamilyList(fontFamily).map(stripOuterQuotes).filter(Boolean);
-  const hasFallback = parts.some(
-    (part) => part.trim().toLowerCase() === TERMINAL_ICON_FALLBACK_FAMILY.toLowerCase()
-  );
-  if (hasFallback) {
-    return formatCssFontFamilyList(fontFamily);
-  }
-
-  const withFallback =
-    parts.length === 0
-      ? TERMINAL_ICON_FALLBACK_FAMILY
-      : `${parts.join(", ")}, ${TERMINAL_ICON_FALLBACK_FAMILY}`;
-
-  return formatCssFontFamilyList(withFallback);
-}
-
-const FONT_AVAILABILITY_TEST_STRING = "abcdefghijklmnopqrstuvwxyz0123456789";
-const FONT_AVAILABILITY_BASE_FAMILIES = ["monospace", "serif", "sans-serif"] as const;
-
-function isFontFamilyAvailableInBrowser(family: string, fontSize: number): boolean {
-  if (typeof document === "undefined") {
-    return true;
-  }
-
-  const body = document.body;
-  if (!body) {
-    return true;
-  }
-
-  const span = document.createElement("span");
-  span.textContent = FONT_AVAILABILITY_TEST_STRING;
-  span.style.position = "absolute";
-  span.style.left = "-9999px";
-  span.style.top = "-9999px";
-  span.style.fontSize = `${fontSize}px`;
-  span.style.fontVariant = "normal";
-  span.style.fontStyle = "normal";
-  span.style.fontWeight = "400";
-  span.style.letterSpacing = "0";
-  span.style.whiteSpace = "nowrap";
-  body.appendChild(span);
-
-  try {
-    const baselineWidths = new Map<string, number>();
-
-    for (const baseFamily of FONT_AVAILABILITY_BASE_FAMILIES) {
-      span.style.fontFamily = baseFamily;
-      baselineWidths.set(baseFamily, span.offsetWidth);
-    }
-
-    for (const baseFamily of FONT_AVAILABILITY_BASE_FAMILIES) {
-      span.style.fontFamily = `${quoteCssFontFamily(family)}, ${baseFamily}`;
-
-      const baseline = baselineWidths.get(baseFamily);
-      if (baseline === undefined) {
-        continue;
-      }
-
-      if (span.offsetWidth !== baseline) {
-        return true;
-      }
-    }
-
-    return false;
-  } finally {
-    span.remove();
-  }
 }
 
 function canLoadFontFamily(primary: string, fontSize: number): boolean {
@@ -226,6 +93,35 @@ function resolveTerminalFontFamily(fontFamily: string, fontSize: number): string
   }
 
   return appendTerminalIconFallback(formatted);
+}
+
+const TERMINAL_FONT_LOAD_TEST_STRING = "abcdefghijklmnopqrstuvwxyz0123456789";
+const TERMINAL_ICON_LOAD_TEST_STRING = String.fromCodePoint(0xf024b);
+
+async function preloadTerminalWebfonts(
+  resolvedFontFamily: string,
+  fontSize: number
+): Promise<void> {
+  if (typeof document === "undefined") {
+    return;
+  }
+
+  const fontFaceSet = document.fonts;
+  if (!fontFaceSet || typeof fontFaceSet.load !== "function") {
+    return;
+  }
+
+  try {
+    await Promise.all([
+      fontFaceSet.load(`${fontSize}px ${resolvedFontFamily}`, TERMINAL_FONT_LOAD_TEST_STRING),
+      fontFaceSet.load(
+        `${fontSize}px ${formatCssFontFamilyList(TERMINAL_ICON_FALLBACK_FAMILY)}`,
+        TERMINAL_ICON_LOAD_TEST_STRING
+      ),
+    ]);
+  } catch (err) {
+    console.warn("[TerminalView] Failed to preload webfonts:", err);
+  }
 }
 
 interface TerminalViewProps {
@@ -503,6 +399,8 @@ export function TerminalView({
           terminalFontConfig.fontSize
         );
 
+        await preloadTerminalWebfonts(resolvedFontFamily, terminalFontConfig.fontSize);
+
         if (cancelled) {
           return;
         }
@@ -678,6 +576,8 @@ export function TerminalView({
         terminalFontConfig.fontFamily,
         terminalFontConfig.fontSize
       );
+
+      await preloadTerminalWebfonts(resolvedFontFamily, terminalFontConfig.fontSize);
 
       if (cancelled || term !== termRef.current) {
         return;
