+++ src/common/constants/storage.ts
@@ -290,7 +290,7 @@ export interface TerminalFontConfig {
 }
 
 export const DEFAULT_TERMINAL_FONT_CONFIG: TerminalFontConfig = {
-  fontFamily: "ui-monospace, monospace",
+  fontFamily: "Geist Mono, ui-monospace, monospace",
   fontSize: 13,
 };
